move iocsh function to init file and global vars to caster struct
