+++ client/castApp/src/reccaster.dbd
@@ -1,5 +1,4 @@
 registrar(reccasterRegistrar)
-registrar(addReccasterEnvVarRegistrar)
 variable(reccastTimeout,double)
 variable(reccastMaxHoldoff,double)
 device(mbbi, INST_IO, devCasterMBBIState, "RecCaster")
